Create core-capture module

+++ feature-scenario-config/src/main/java/com/buzbuz/smartautoclicker/feature/scenario/config/ui/scenario/config/ScenarioConfigViewModel.kt
@@ -26,12 +26,12 @@ import androidx.lifecycle.viewModelScope
 import com.buzbuz.smartautoclicker.core.ui.bindings.DropdownItem
 import com.buzbuz.smartautoclicker.feature.billing.IBillingRepository
 import com.buzbuz.smartautoclicker.feature.billing.ProModeAdvantage
-import com.buzbuz.smartautoclicker.detection.DETECTION_QUALITY_MAX
-import com.buzbuz.smartautoclicker.detection.DETECTION_QUALITY_MIN
 import com.buzbuz.smartautoclicker.feature.scenario.config.domain.EditionRepository
 import com.buzbuz.smartautoclicker.domain.model.AND
 import com.buzbuz.smartautoclicker.domain.model.OR
 import com.buzbuz.smartautoclicker.domain.model.endcondition.EndCondition
+import com.buzbuz.smartautoclicker.domain.model.scenario.DETECTION_QUALITY_MAX
+import com.buzbuz.smartautoclicker.domain.model.scenario.DETECTION_QUALITY_MIN
 import com.buzbuz.smartautoclicker.feature.scenario.config.R
 
 import kotlinx.coroutines.flow.combine
@@ -180,7 +180,7 @@ class ScenarioConfigViewModel(application: Application) : AndroidViewModel(appli
             viewModelScope.launch {
                 editionRepository.updateEditedScenario(
                     scenario.copy(
-                        detectionQuality = max(scenario.detectionQuality - 1, DETECTION_QUALITY_MIN.toInt())
+                        detectionQuality = max(scenario.detectionQuality - 1, DETECTION_QUALITY_MIN)
                     )
                 )
             }
@@ -193,7 +193,7 @@ class ScenarioConfigViewModel(application: Application) : AndroidViewModel(appli
             viewModelScope.launch {
                 editionRepository.updateEditedScenario(
                     scenario.copy(
-                        detectionQuality = min(scenario.detectionQuality + 1, DETECTION_QUALITY_MAX.toInt())
+                        detectionQuality = min(scenario.detectionQuality + 1, DETECTION_QUALITY_MAX)
                     )
                 )
             }
